# vulkan: further optimize q5_k mul_mat_vec