@daniandtheweb 2024-11-26
These changes make quite a big difference on my Radeon 5700XT.

| model                          |       size |     params | backend    | ngl | threads |          test |         branch |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ------------: | ------------------: | -------------------: |
| qwen2 7B Q5_K - Small          |   4.94 GiB |     7.62 B | Vulkan     |  99 |       4 |         tg128 |   master |     41.07 ± 0.06 |
| qwen2 7B Q5_K - Small          |   4.94 GiB |     7.62 B | Vulkan     |  99 |       4 |         tg128 |    PR |     49.23 ± 0.42 |