@netrunnereve 2024-11-26
I haven't tried it with an actual model but our tests show that it's now 6% faster on a RX 570.

```
Master:
  MUL_MAT(type_a=q5_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   1704 runs -  1195.05 us/run - 117.44 MFLOP/run -  98.27 GFLOPS
PR:
  MUL_MAT(type_a=q5_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   1704 runs -  1124.40 us/run - 117.44 MFLOP/run - 104.45 GFLOPS
```